requirements.txt
phidata
agno
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai